chore(heureka): adds SeverityCount component as a separate file which is imported by many others
